use only one GitHubService
