+++ Client/GitHubCardClientStartup.cs
@@ -9,8 +9,7 @@ namespace RyanJagdfeld.Module.GitHubCard.Startup
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddScoped<IGitHubCardService, GitHubCardService>();
-            services.AddScoped(implementationFactory => new GitHubRepoService(new System.Net.Http.HttpClient())); 
-            services.AddScoped(implementationFactory => new GitHubUserService(new System.Net.Http.HttpClient())); 
+            services.AddScoped(implementationFactory => new GitHubService(new System.Net.Http.HttpClient()));
         }
     }
 }
